Break down userProfile in multiple files
