--- src/components/homePage/contexts.js
@@ -1,1 +1,0 @@
-
